LPS-192496 Move clusterConfigurationModelIndexer.initialize(); from SearchResultsMVCRenderCommand to ConfigurationModelIndexer#search, Move initialize() from ClusterConfigurationModelIndexer to ConfigurationModelIndexer and prepare for next
